+++ distributed_nose/__init__.py
@@ -1,5 +1,5 @@
 """Distribute your nose tests across multiple machines, hassle-free"""
-VERSION = (0, 1, 3, '')
+VERSION = (0, 2, 0, '')
 __version__ = '.'.join(map(str, VERSION[0:3])) + ''.join(VERSION[3:])
 __author__ = 'Wes Winham'
 __contact__ = 'winhamwr@gmail.com'